+++ test/opf.jl
@@ -155,6 +155,44 @@
             @test result["termination_status"] == LOCALLY_SOLVED
             @test isapprox(result["objective"], 0.054; atol=1e-4)
         end
+
+        @testset "test warm start from dss voltages export" begin
+            eng = deepcopy(case5_phase_drop)
+            add_voltage_starts!(eng, "../test/data/opendss/case5_voltage.csv")
+            apply_voltage_bounds!(eng)
+
+            for (_,bus) in eng["bus"]
+                nt = length(bus["terminals"])
+                @test haskey(bus, "vm_start") && length(bus["vm_start"]) == nt
+                @test haskey(bus, "va_start") && length(bus["va_start"]) == nt
+            end
+
+            math = transform_data_model(eng)
+
+            for (_,bus) in math["bus"]
+                if startswith(bus["source_id"], "bus")
+                    @test bus["vm_start"] <= bus["vmax"] && bus["vm_start"] >= bus["vmin"]
+                end
+            end
+
+            pm = instantiate_mc_model(math, ACPUPowerModel, build_mc_opf)
+            for (i,vm) in var(pm, nw_id_default, :vm)
+                if startswith(math["bus"]["$i"]["source_id"], "bus")
+                    @test all(JuMP.start_value(vm[t]) == math["bus"]["$i"]["vm_start"][idx] for (idx,t) in enumerate(math["bus"]["$i"]["terminals"]))
+                end
+            end
+
+            for (i,va) in var(pm, nw_id_default, :va)
+                if startswith(math["bus"]["$i"]["source_id"], "bus")
+                    @test all(JuMP.start_value(va[t]) == math["bus"]["$i"]["va_start"][idx] for (idx,t) in enumerate(math["bus"]["$i"]["terminals"]))
+                end
+            end
+
+            result = solve_mc_opf(eng, ACRUPowerModel, ipopt_solver; solution_processors=[sol_data_model!], make_si=false)
+
+            @test all(isapprox.(result["solution"]["bus"]["midbus"]["vm"], [0.97352, 0.9649, 0.95646]; atol=1e-4))
+            @test all(isapprox.(result["solution"]["bus"]["midbus"]["va"], [-1.3, -121.3, 118.2]; atol=1e-1))
+        end
     end
 
     @testset "test opendss opf" begin