+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## staged
 
+- Added support for parsing the results of dss command `export voltages` into `vm_start` and `va_start` values for warm starting bus voltages
 - Added support for raw parsing of RegControl and CapControl objects from dss
 - Added delta/voltage-dependent loads to LinDist3Flow formulation